Add files via upload
Before using this function, make sure the uploaded file is stored in an array. This function will take both the file array, and common words array as inputs. It will run matching, remove any matches founded. Then it will return the result in an array. All the words are in lower case.